Ship communicates with MainApp now, moved game loop to MainApp
This was a lot of trial and error, but I finally figured out where the
main game loop is supposed to be. SomePane has the mouse listeners,
MainApp has the game loop, and I have some test code in the game loop to
begin with. Also, Ship class now communicates with MainApp so that it
can see other ships in the game and add projectiles to the array
directly from its shoot() function. Basically makes the Ship class much
more versatile.